requirements.txt
typer
rich
panda
httpx
tiktoken